refactor qt gui placement in code
