+++ gui/qt/README.md
