+++ gui/qt/depthai_demo.pyproject
