+++ gui/qt/views/AIProperties.qml
