+++ gui/qt/views/CameraPreview.qml
