+++ gui/qt/views/CameraProperties.qml
