+++ gui/qt/views/DepthProperties.qml
