+++ gui/qt/views/MiscProperties.qml
